+++ server/extension/cdb_dataservices_server--0.39.0--0.38.0.sql
@@ -1,4 +1,9 @@
--- PG12_DEPRECATED
+--DO NOT MODIFY THIS FILE, IT IS GENERATED AUTOMATICALLY FROM SOURCES
+-- Complain if script is sourced in psql, rather than via CREATE EXTENSION
+\echo Use "ALTER EXTENSION cdb_dataservices_server UPDATE TO '0.38.0'" to load this file. \quit
+
+-- HERE goes your code to upgrade/downgrade
+
 -- Create geomval if it doesn't exist (in postgis 3+ it only exists in postgis_raster)
 DO $$
 BEGIN